refactor: extract 4 hooks from AppLayout (776→462 lines)
AppLayout.tsx was a 776-line god component with 7 responsibilities.
Extract business logic into focused custom hooks:

- useTeamSync (209 lines) — API restoration, WS connection, sync:init,
  REST fallback, data normalization, auto-select first team
- usePresenceEvents (84 lines) — presence + voice occupant WS subscriptions
- useIdentityBackup (50 lines) — identity blob upload to servers
- useCryptoRestore (29 lines) — CryptoManager re-init from derivedKey

AppLayout now handles only: layout rendering, view mode state, mobile
tabs, modals, and keyboard shortcuts. Each hook is self-contained with
its own store/service imports and ref guards.
